+++ src/main/java/by/siarhei/pikabu/math/AmperageCalculator.java
@@ -2,7 +2,7 @@ package by.siarhei.pikabu.math;
 
 public class AmperageCalculator implements BaseCalculator {
     @Override
-    public double calculate(double diameter, double length,int amper) {
+    public double calculate(double diameter, double length,double amper) {
         return ((Math.PI * diameter * (length + diameter / 2)) / 10000) * amper;
     }
 }
