+++ src/main/java/by/siarhei/pikabu/math/BaseCalculator.java
@@ -1,5 +1,5 @@
 package by.siarhei.pikabu.math;
 
 public interface BaseCalculator {
-    double calculate(double diameter, double length,int amper);
+    double calculate(double diameter, double length,double amper);
 }
